Cleaned up errors, unnecessary code, unstructured index files, and function component states
