--- client/src/components/Navbar/Menus/Unauthenticated/Mobile.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const Mobile = () => {
-  return <div></div>;
-};
-
-export default Mobile;
